Extract base class for LDAP JDBC tests
